requirements.txt
## Stdlib
# math
##
Pillow
numpy